+++ android/app/src/main/java/edu/illinois/rokwire/MainActivity.java
@@ -41,7 +41,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import com.rokmetro.university.neom.mobile_access.MobileAccessPlugin;
+//import com.rokmetro.university.neom.mobile_access.MobileAccessPlugin;
 import io.flutter.embedding.android.FlutterActivity;
 import io.flutter.embedding.engine.FlutterEngine;
 import io.flutter.plugin.common.MethodCall;
@@ -62,24 +62,24 @@ public class MainActivity extends FlutterActivity implements MethodChannel.Metho
 
     private Toast statusToast;
 
-    private MobileAccessPlugin mobileAccessPlugin;
+//    private MobileAccessPlugin mobileAccessPlugin;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
         initScreenOrientation();
-        if (mobileAccessPlugin != null) {
-            mobileAccessPlugin.onActivityCreate();
-        }
+//        if (mobileAccessPlugin != null) {
+//            mobileAccessPlugin.onActivityCreate();
+//        }
     }
 
     @Override
     protected void onStart() {
         super.onStart();
-        if (mobileAccessPlugin != null) {
-            mobileAccessPlugin.onActivityStart();
-        }
+//        if (mobileAccessPlugin != null) {
+//            mobileAccessPlugin.onActivityStart();
+//        }
     }
 
     @Override
@@ -89,9 +89,9 @@ public class MainActivity extends FlutterActivity implements MethodChannel.Metho
         if (orientationListener != null) {
             orientationListener.disable();
         }
-        if (mobileAccessPlugin != null) {
-            mobileAccessPlugin.onActivityDestroy();
-        }
+//        if (mobileAccessPlugin != null) {
+//            mobileAccessPlugin.onActivityDestroy();
+//        }
     }
 
     @Override
@@ -105,8 +105,8 @@ public class MainActivity extends FlutterActivity implements MethodChannel.Metho
         METHOD_CHANNEL = new MethodChannel(flutterEngine.getDartExecutor().getBinaryMessenger(), NATIVE_CHANNEL);
         METHOD_CHANNEL.setMethodCallHandler(this);
 
-        mobileAccessPlugin = new MobileAccessPlugin(this);
-        flutterEngine.getPlugins().add(mobileAccessPlugin);
+//        mobileAccessPlugin = new MobileAccessPlugin(this);
+//        flutterEngine.getPlugins().add(mobileAccessPlugin);
     }
 
     private void initScreenOrientation() {