+++ lib/service/Config.dart
@@ -260,6 +260,9 @@ class Config extends rokwire.Config {
   bool get enableLogging => false;
 
   @override
+  String? get packageName => kIsWeb ? 'neom' : super.packageName;
+
+  @override
   String? get appId => kIsWeb ? 'com.rokmetro.university.${super.appId}' : super.appId;
 
   @override