+++ lib/ui/groups/GroupPendingMemberPanel.dart
@@ -55,6 +55,13 @@ class _GroupPendingMemberPanelState extends State<GroupPendingMemberPanel> {
   bool _updating = false;
 
   @override
+  void initState() {
+    _denied = widget.member?.isRejected == true;
+    _approved = widget.member?.isMemberOrAdmin == true;
+    super.initState();
+  }
+
+  @override
   void dispose() {
     _reasonController.dispose();
     super.dispose();
@@ -210,7 +217,7 @@ class _GroupPendingMemberPanelState extends State<GroupPendingMemberPanel> {
                       child: TextField(
                         controller: _reasonController,
                         decoration: InputDecoration(border: InputBorder.none),
-                        style: Styles().textStyles.getTextStyle("widget.title.regular"),
+                        style: Styles().textStyles.getTextStyle("widget.title.dark.regular"),
                         onChanged: (text){setState(() {});},
                         minLines: 4,
                         maxLines: 999,
@@ -243,8 +250,10 @@ class _GroupPendingMemberPanelState extends State<GroupPendingMemberPanel> {
                 padding: EdgeInsets.symmetric(horizontal: 16, vertical: 13),
                 progress: _updating,
                 onTap: () {
-                  Analytics().logSelect(target: 'Apply');
-                  _processMembership();
+                  if(_approved || (_denied && _reasonController.text.isNotEmpty)){
+                    Analytics().logSelect(target: 'Apply');
+                    _processMembership();
+                  }
                 },
               ),
             )
@@ -261,6 +270,7 @@ class _GroupPendingMemberPanelState extends State<GroupPendingMemberPanel> {
       _updating = true;
     });
 
+    widget.member?.status = _approved ? GroupMemberStatus.member : GroupMemberStatus.rejected;
     Groups().acceptMembership(widget.group, widget.member, _approved, _reasonController.text).then((bool result) {
       if (mounted) {
         setState(() {