+++ lib/ui/profile/ProfileInfoEditPage.dart
@@ -36,14 +36,13 @@ class ProfileInfoEditPage extends StatefulWidget {
 
   final Uint8List? pronunciationAudioData;
   final Uint8List? photoImageData;
-  final String? photoImageToken;
 
-  final void Function({Auth2UserProfile? profile, Auth2UserPrivacy? privacy, Uint8List? pronunciationAudioData, Uint8List? photoImageData, String? photoImageToken})? onFinishEdit;
+  final void Function({Auth2UserProfile? profile, Auth2UserPrivacy? privacy, Uint8List? pronunciationAudioData, Uint8List? photoImageData})? onFinishEdit;
 
   ProfileInfoEditPage({super.key,
     this.onboarding = false,
     this.authType, this.profile, this.privacy, this.identifiers,
-    this.pronunciationAudioData, this.photoImageData, this.photoImageToken,
+    this.pronunciationAudioData, this.photoImageData,
     this.onFinishEdit
   });
 
@@ -56,7 +55,6 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
   late Auth2UserProfileFieldsVisibility _profileVisibility;
   late Uint8List? _pronunciationAudioData;
   late Uint8List? _photoImageData;
-  late String? _photoImageToken;
 
   final Map<_ProfileField, Auth2FieldVisibility?> _fieldVisibilities = {};
   final Map<_ProfileField, TextEditingController> _fieldTextControllers = {};
@@ -132,7 +130,6 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
 
     _pronunciationAudioData = widget.pronunciationAudioData;
     _photoImageData = widget.photoImageData;
-    _photoImageToken = widget.photoImageToken;
     _identifiers = List.generate(widget.identifiers?.length ?? 0, (index) => Auth2PublicAccountIdentifier.fromUserIdentifier(widget.identifiers![index]));
 
     for (_ProfileField field in _ProfileField.values) {
@@ -252,19 +249,14 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
 
     // Edit: Photo
 
-    String? get _photoImageUrl => StringUtils.isNotEmpty(_photoText) ?
-      Content().getUserPhotoUrl(type: UserProfileImageType.medium, params: DirectoryProfilePhotoUtils.tokenUrlParam(_photoImageToken)) : null;
-
     double get _photoImageSize => MediaQuery.of(context).size.width / 3;
-
-    Map<String, String>? get _photoAuthHeaders => DirectoryProfilePhotoUtils.authHeaders;
 
     Widget get _photoWidget => Stack(children: [
       Padding(padding: EdgeInsets.only(left: 16, right: 16, bottom: 20), child:
         DirectoryProfilePhoto(
           key: _photoKey,
-          photoUrl: _photoImageUrl,
-          photoUrlHeaders: _photoAuthHeaders,
+          accountId: StringUtils.isNotEmpty(_photoText) ? Auth2().accountId : null,
+          type: UserProfileImageType.medium,
           photoData: _photoImageData,
           imageSize: _photoImageSize,
         ),
@@ -320,7 +312,6 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
             setState(() {
               _photoKey = UniqueKey();
               _photoText = Content().getUserPhotoUrl(accountId: Auth2().accountId, type: UserProfileImageType.medium) ?? '';
-              _photoImageToken = DirectoryProfilePhotoUtils.newToken;
               _photoImageData = imageUploadResult.imageData;
             });
           }
@@ -359,7 +350,6 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
             setState(() {
               _photoKey = UniqueKey();
               _photoText = '';
-              _photoImageToken = DirectoryProfilePhotoUtils.newToken;
               _photoImageData = null;
             });
           }
@@ -477,7 +467,7 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
   }
 
   void _onEditPronunciation() {
-    Analytics().logSelect(target: 'Edit Pronuncaion');
+    Analytics().logSelect(target: 'Edit Pronunciation');
     _createPronunciation();
   }
 
@@ -485,7 +475,7 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
     ProfileSoundRecorderDialog.show(context).then((AudioResult? result) {
       if (result?.resultType == AudioResultType.succeeded) {
         setState(() {
-          _pronunciationText = Content().getUserNamePronunciationUrl(accountId: Auth2().accountId);
+          _pronunciationText = Content().getUserNamePronunciationFileName(accountId: Auth2().accountId);
           _pronunciationAudioData = result?.audioData;
         });
       }
@@ -493,30 +483,47 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
   }
 
   void _onDeletePronunciation() {
-    Analytics().logSelect(target: 'Delete Pronuncaion');
+    Analytics().logSelect(target: 'Delete Pronunciation');
 
     AppAlert.showConfirmationDialog(context, message: Localization().getStringEx("panel.profile_info.pronunciation.delete.confirmation.msg", "Are you sure you want to remove this pronunciation audio?")).then((bool? result) {
       if (mounted && (result == true)) {
         setState(() {
           _clearingUserPronunciation = true;
         });
-        Content().deleteUserNamePronunciation().then((AudioResult? result){
-          if (mounted) {
-            if (result?.resultType == AudioResultType.succeeded) {
-              setState(() {
-                _clearingUserPronunciation = false;
-                _pronunciationText = null;
-                _pronunciationAudioData = null;
-              });
+        List<String> splitPronunciationText = _pronunciationText?.split('.') ?? [];
+        if (splitPronunciationText.length >= 2) {
+          Content().deleteUserNamePronunciation(extension: '.${splitPronunciationText.last}').then((AudioResult? result){
+            if (mounted) {
+              if (result?.resultType == AudioResultType.succeeded) {
+                setState(() {
+                  _clearingUserPronunciation = false;
+                  _pronunciationText = null;
+                  _pronunciationAudioData = null;
+                  if (widget.privacy != null) {
+                    Auth2UserProfile profile = _Auth2UserProfileUtils.buildModified(widget.profile, _fieldTextControllers);
+                    Auth2UserPrivacy privacy = Auth2UserPrivacy.fromOther(widget.privacy);
+                    _saveEdit(profile, privacy).then((result) {
+                      if (result.succeeded) {
+                        widget.onFinishEdit?.call(
+                          profile: (result.profile == true) ? profile : null,
+                          privacy: (result.privacy == true) ? privacy : null,
+                          pronunciationAudioData: _pronunciationAudioData,
+                          photoImageData: _photoImageData,
+                        );
+                      }
+                    });
+                  }
+                });
+              }
+              else {
+                setState(() {
+                  _clearingUserPronunciation = false;
+                });
+                AppAlert.showTextMessage(context, Localization().getStringEx('panel.profile_info.pronunciation.delete.failed.msg', 'Failed to delete pronunciation audio. Please try again later.'));
+              }
             }
-            else {
-              setState(() {
-                _clearingUserPronunciation = false;
-              });
-              AppAlert.showTextMessage(context, Localization().getStringEx('panel.profile_info.pronunciation.delete.failed.msg', 'Failed to delete pronunciation audio. Please try again later.'));
-            }
-          }
-        });
+          });
+        }
       }
     });
   }
@@ -530,7 +537,7 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
 
         Uint8List? audioData = _pronunciationAudioData;
         if (audioData == null) {
-          AudioResult? result = await Content().loadUserNamePronunciation();
+          AudioResult? result = StringUtils.isNotEmpty(_pronunciationText) ? await Content().loadUserNamePronunciation(fileName: _pronunciationText) : null;
           audioData = (result?.resultType == AudioResultType.succeeded) ? result?.audioData : null;
         }
 
@@ -1115,14 +1122,12 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
             privacy: (result.privacy == true) ? privacy : null,
             pronunciationAudioData: _pronunciationAudioData,
             photoImageData: _photoImageData,
-            photoImageToken: _photoImageToken,
           );
         }
       }
       else if (shouldSave == false) {
         widget.onFinishEdit?.call(
           photoImageData: _photoImageData,
-          photoImageToken: _photoImageToken,
           pronunciationAudioData: _pronunciationAudioData,
         );
       }
@@ -1202,7 +1207,6 @@ class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with Notificat
           privacy: (result.privacy == true) ? privacy : null,
           pronunciationAudioData: _pronunciationAudioData,
           photoImageData: _photoImageData,
-          photoImageToken: _photoImageToken,
         );
       }
     }