+++ CHANGELOG.md
@@ -5,8 +5,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - NEOM U
+## [0.5.17] = 2025-05-10
+### Changed
+- Update media file transfers to use AWS S3 directly [#104](https://github.com/rokmetro/neom-app/issues/104)
+### Fixed
+- Slow image/video/file display and upload for conversations and groups [#91](https://github.com/rokmetro/neom-app/issues/91)
+- Error saving business card QR code[#90](https://github.com/rokmetro/neom-app/issues/90)
+- Fixed package name bug for mobile platforms
+
+## [0.5.16] = 2025-05-02
+### Added
+- Change group membership decision after rejecting [#102](https://github.com/rokmetro/neom-app/issues/102)
 ### Fixed
 - Visual Cleanup [#83](https://github.com/rokmetro/neom-app/issues/83)
+- Notifications not received [#92](https://github.com/rokmetro/neom-app/issues/92)
+- Group UI Fixes [#89](https://github.com/rokmetro/neom-app/issues/89)
+- Remove the option to add past events to groups [#87](https://github.com/rokmetro/neom-app/issues/87)
+### Changed
+- Allow attachments with messages in conversations [#85](https://github.com/rokmetro/neom-app/issues/85)
 
 ## [0.5.15] = 2025-03-20
 ### Changed